Move seqpriorityqueue to gateway
